Watchers may get closed after server restart if db was compacted
Every watcher internally remembers the revision of the last
notification it has received for the watched prefix.
The revision currently in use by the server may be higher
if keys outside of the watched scope were edited since then.
When etcd server restarts, watcher will ask for all changes since
the last revision it has seen. This is not a problem even if the
revision is older than the (restored) current one, provided that
the server has kept the history of changes in-between. However,
this is not the case if etcd was compacted between the the last
notification delivered to the watcher and the server restart.
In such case, the server refuses the watcher to reconnect after
the restart, with a reason that it can only deliver notifications
since the last compaction.
This is not a fatal case, however, since changes between the last
delivered notification and the compaction were not related to the
watched prefix. Watcher just needs to re-create subscription,
starting with the revision of the compaction.

Signed-off-by: Milan Lenco <milan.lenco@pantheon.tech>