Added a wiring package to make it easy to wire plugin dependencies
Wiring packages with Flavors is hard and confusing.  It also encourages
overinclusion, because plugins come in big bundles some of which you
need and some of which you don't.

Wiring is a simple concept.  A plugin is Wireable if you can call:

plugin.Wire(wiring)

Wiring is just a function:

type Wiring func (core.Plugin) error

that is expected to wire up the plugins dependencies.

A plugin is said to be DefaultWireable if it implements:

func DefaultWiring() Wiring

so you can call

plugin.Wire(plugin.DefaultWiring())

conventionally, this is equivalent to plugin.Wire(nil)

Helper functions are provided to let you quickly get from a Wireable Plugin
to an Agent or running EventLoop

agent,err :=wiring.NewAgent(plugin)

err := wiring.EventLoop(plugin)

The net result of all of this is you can in one or two lines have a fully
functioning EventLoop running for your plugin.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>