Make kvdbsync Init() and AfterInit() idempotent
... and also threadsafe.  Used sync.Once to do so.

This is complicated because of the binding of
OfDifferentAgent to a non-pointer version of Plugin.

Because OfDifferentAgent copies the entire Plugin
struct, including its Once's.  Once's contain
mutexes.  go vet screams bloody murder (rightly)
when you copy mutexes.

So I bound OfDifferentAgent to a pointer to the
original Plugin, created a new Plugin, and manually
copied the Deps.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>