Signals are hard :(
After much work to try to test signals, I realized a few
things:

1) If we don't set ourselves up to be Notified *before*
we try to start the agent, we can get lots of fun
race conditions due to signals received *while* the
agent is starting.  That turns out to be bad.
2) Its much simpler to *wrap* the logic to handle
signals *around* start() than to insert it into
start()

So I did.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>