Make grpc config accessible from outside grpc package
Currently, the grpc.grpcCfg is inaccessible from outside
the grpc package, making it impossible to set up a config
for a GRPC plugin except from config files.

This patch follows the pattern of other plugins of renaming
it
grpc.Config
to allow for it to be set other than from config files.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>